requirements.txt
httpx
pandas
SQLAlchemy
python-dotenv
psycopg2